requirements.txt
#### Bot lib ####
git+https://gitlab.com/HarukaNetwork/OSS/PTB.git
telethon
#################
aiohttp>=2.2.5
bleach
feedparser
hurry.filesize
markdown2
psycopg2-binary
PyLyrics
python-rapidjson
PyYAML
requests>=2.20.0
spamwatch
spongemock
tldextract
covid
future
emoji
beautifulsoup4
sqlalchemy
python-telegram-bot==11.1.0
pynewtonmath
zalgo-text
geopy
nltk
psutil
Pillow>=4.2.0
CurrencyConverter
googletrans
jikanpy
speedtest-cli
coffeehouse
tswift
gTTS
wikipedia
pyowm
tabulate
parsel
PyPDF2
qrcode
aiogram